Refactor example and tutorial
