+++ Sources/ArcGISToolkit/Documentation.docc/Resources/Authenticator/AuthenticatorStep4.swift
@@ -1,24 +1,6 @@
-import SwiftUI
-import ArcGISToolkit
-import ArcGIS
-
-@main
-struct AuthenticationApp: App {
-    @ObservedObject var authenticator: Authenticator
-    
-    init() {
-        authenticator = Authenticator()
-        ArcGISEnvironment.authenticationManager.handleChallenges(using: authenticator)
-    }
-    
-    var body: some SwiftUI.Scene {
-        WindowGroup {
-            HomeView()
-                .authenticator(authenticator)
-                .environmentObject(authenticator)
-                .task {
-                    try? await ArcGISEnvironment.authenticationManager.setupPersistentCredentialStorage(access: .whenUnlockedThisDeviceOnly)
-                }
-        }
+func signOut() {
+    Task {
+        await ArcGISEnvironment.authenticationManager.revokeOAuthTokens()
+        await ArcGISEnvironment.authenticationManager.clearCredentialStores()
     }
 }
