--- Sources/ArcGISToolkit/Documentation.docc/Resources/Authenticator/AuthenticatorStep5.swift
@@ -1,6 +1,0 @@
-func signOut() {
-    Task {
-        await ArcGISEnvironment.authenticationManager.revokeOAuthTokens()
-        await ArcGISEnvironment.authenticationManager.clearCredentialStores()
-    }
-}
