+++ Sources/ArcGISToolkit/Documentation.docc/Tutorials/AuthenticatorTutorial.tutorial
@@ -13,27 +13,25 @@
         @Steps {
             @Step {
                 Declare and initialize an `Authenticator` member in your app.
+                
+                For OAuth implementations use `Authenticator(oAuthUserConfigurations:)`.
                 @Code(name: "ExampleApp.swift", file: AuthenticatorStep1)
             }
             
             @Step {
-                Set the `Authenticator` to handle challenges.
+                Declare the app's body, placing the `Authenticator` into the hierarchy.
                 @Code(name: "ExampleApp.swift", file: AuthenticatorStep2)
             }
             
             @Step {
-                Declare the app's body, placing the `Authenticator` into the hierarchy.
+                Set the `Authenticator` to handle challenges.
+                Set up credential persistence.
                 @Code(name: "ExampleApp.swift", file: AuthenticatorStep3)
             }
             
             @Step {
-                Setup credential persistence.
-                @Code(name: "ExampleApp.swift", file: AuthenticatorStep4)
-            }
-            
-            @Step {
                 During sign-out, use these `AuthenticationManager` methods.
-                @Code(name: "SignOutView.swift", file: AuthenticatorStep5)
+                @Code(name: "SignOutView.swift", file: AuthenticatorStep4)
             }
         }
     }
